requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
pillow
numpy
tensorflow
langchain
langgraph
langchain-openai
openai
redis
